🐛 fix : rename tab
